requirements.txt
langchain
langchain-openai
langchain-groq
streamlit
python-dotenv
mysql-connector-python
sqlmodel